requirements.txt
pandas
japanize-matplotlib
jquants-api-client
pyyaml
snowflake-snowpark-python
snowflake-connector-python
python-dateutil